+++ CHANGELOG.md
@@ -1,7 +1,56 @@
-# [Unreleased] - 2024-11-11
+# [Unreleased-main] - 2024-12-12
 
 ## Added
 - Default database for gas pipe dimensions based on the ASA pipe schedule with thicknesses from the standard class
+- Gas and electricity workflow has been added. Still a work in progress
+- Internal energy content function added for natural gas and hydrogen
+- OPEX added for gas and electricity source
+- Added function to make pickle usable (exceptions) in the OMOTES back end
+- More of the existing classes added to __init_ for local runs using the grow_workflow
+
+## Changed
+- Write output to json also saves the solver statistics.
+- Updates in asset conversion from ESDL to pycml; gas and heat pipe split, generic conversion better specified.
+- Adapt profiles for heating and cooling peak day seperately & add test case
+- For gas: Use energy content (heating value) instead of internal energy 
+- Option was added to allow a bypass of the heat exchanger 
+
+## Fixed
+- Bugfix: gas boiler mass flow contraint units
+- Bugfix: same mip gap settings for all solvers in grow_workflow.
+- Bugfix: head loss test case when minimum_velocity = 0.0
+- Bugs: state update of heat pump, heat buffer volume & ates charge rates update in ESDL, heat storage asset data output to influxDB  
+- Bugfix: get_density + get_internal_energy updated to ensure it checks commodity type for heat instead of the carrier name
+
+
+# [0.1.8.4] - 2024-12-11
+
+## Added
+- More of the existing classes added to __init_ for local runs using the grow_workflow
+
+## Changed
+- xxx
+
+## Fixed
+- Bugs: state update of heat pump, heat buffer volume & ates charge rates update in ESDL, heat storage asset data output to influxDB   
+
+
+# [0.1.8.3] - 2024-12-02
+
+## Added
+- Added function to make pickle usable (exceptions) in the OMOTES back end 
+
+## Changed
+- xx
+
+## Fixed
+- xx
+
+
+# [0.1.8.2] - 2024-11-15
+
+## Added
+- xx
 
 ## Changed
 - xx
@@ -11,7 +60,7 @@
 - Bugfix: head loss test case when minimum_velocity = 0.0
 
 
-# [0.1.8] - 2024-11-07
+# [0.1.8 & 0.1.8.1] - 2024-11-07
 
 ## Added
 - Added MESIDO error exception class to raise applicable potential errors in a workflow & include required info for feedback in the mapeditor